EventsWorker: Send HTTP results to the worker.
Due to the nature of MQTT, the HTTP responses are not sent to the
worker(k4e-device-worker), so the worker cannot take decisions based on
the operator.

Like, renew keys to login into the operator. So a worker can check the
401 results to renew certificates.

With this change, the worker will have an Events function, that can
retrieve the Event information with the following information:

```
message APIResponse {
  int64 status_code = 1;

  string body = 2;

  string directive = 3;

  map<string, string> metadata = 4;
}
```

The events are triggered by the `transport.PublishReceivedData` to a
channel that it's async.

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>